feat: add error method
