+++ example/android/app/src/main/java/pagopa/ioreactnativeproximity/example/MainActivity.kt
@@ -20,11 +20,4 @@ class MainActivity : ReactActivity() {
    */
   override fun createReactActivityDelegate(): ReactActivityDelegate =
       DefaultReactActivityDelegate(this, mainComponentName, fabricEnabled)
-
-  /**
-   * We need to override the [onCreate] method for react-native-screens to work properly.
-   */
-  override fun onCreate(savedInstanceState: Bundle?) {
-    super.onCreate(null)
-  }
 }
