+++ example/package.json
@@ -11,14 +11,11 @@
   },
   "dependencies": {
     "@pagopa/io-react-native-crypto": "^1.0.1",
-    "@react-navigation/native": "^7.0.14",
-    "@react-navigation/native-stack": "^7.2.0",
     "react": "18.3.1",
     "react-native": "0.76.7",
     "react-native-permissions": "^5.2.5",
     "react-native-qrcode-svg": "^6.3.14",
     "react-native-safe-area-context": "^5.4.1",
-    "react-native-screens": "^4.10.0",
     "react-native-svg": "^15.11.1"
   },
   "devDependencies": {
